Calc.kt

+++ Calc.kt
@@ -38,10 +38,10 @@ fun main() {
     println("Por favor digite o segundo número que você deseja " + TextOperation + ":")
     val n2 = readLine()!!.toInt()
     var total = when (operation){
-        1 -> n1+n2
-        2 -> n1-n2
-        3 -> n1*n2
-        4 -> n1/n2
+        1 -> n1.toFloat()+n2.toFloat()
+        2 -> n1.toFloat()-n2.toFloat()
+        3 -> n1.toFloat()*n2.toFloat()
+        4 -> n1.toFloat()/n2.toFloat()
         else -> ""
     }
     println("A "+ numberToText + " entre " + n1 + operation2 + n2 + " é de: " + total)
